requirements.txt
langchain
langchain-core
langchain-community 
langchain-aws 
boto3 
python-dotenv
